requirements.txt
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.3.1
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0
sqlparse==0.5.3
tzdata==2025.2
